refactor(entity-selection): split particle-tab into list and grid sub-components
Decompose `particle-tab.svelte` into a thin orchestrator plus
`particle-list-view.svelte` and `particle-grid-view.svelte`, with shared
pure helpers (`Particle`, `NAMED_IDS`, `isExternal`, `atomicNumber`,
`isNamed`, `periodicPosition`) extracted to `particle-tab-helpers.ts`.

The orchestrator keeps view-mode state, filter/match logic, the summary
bar, the view toggle, the multi-select bookkeeping, and the picker-level
keyboard navigation registration; both sub-components receive already-
derived arrays plus per-particle callbacks via Svelte 5 `$props()`.

All `data-testid`, `aria-selected`, `aria-pressed`, and `data-available`
contracts are preserved verbatim so existing Playwright E2E selectors
(`picker-particle-grid`, `picker-particle-list`,
`picker-particle-tile-*`, etc.) keep working unchanged.

Closes #615
