+++ CHANGELOG-AI.md
@@ -17,6 +17,9 @@ Use one bullet per session (newest first):
 
 ## Entries (newest first)
 
+- 2026-05-26 — **Stage 5 — entity-selection**: Decompose `particle-tab.svelte` into focused `particle-list-view` and `particle-grid-view` sub-components (#615) (Claude Opus 4.7 via Claude Code)
+  - **Log:** [log](docs/ai-logs/2026-05-26-issue-615-particle-tab-decompose.md)
+
 - 2026-05-26 — **Stage 5 — entity-selection**: Address PR #614 review-thread fixes for particle grid, URL `across` canonicalization, and E2E stability (GPT-5.5 via GitHub Copilot coding agent)
   - **Log:** [log](docs/ai-logs/2026-05-26-pr614-review-thread-fixes.md)
