+++ src/lib/components/entity-selection/particle-tab.svelte
@@ -1,14 +1,14 @@
 <script lang="ts">
   import { cn } from "$lib/utils.js";
   import type { EntitySelectionState } from "$lib/state/entity-selection.svelte";
-  import type { ParticleEntity } from "$lib/wasm/types";
   import type { ExternalOnlyParticle } from "$lib/state/external-compatibility";
   import { ELECTRON_ID } from "$lib/state/entity-selection.svelte";
   import { getParticleListLabel, getParticleSearchText } from "$lib/utils/particle-label";
   import PickerSummaryBar from "./picker-summary-bar.svelte";
+  import ParticleListView from "./particle-list-view.svelte";
+  import ParticleGridView from "./particle-grid-view.svelte";
   import { isAdvancedMode } from "$lib/state/advanced-mode.svelte";
-
-  type Particle = ParticleEntity | ExternalOnlyParticle;
+  import { atomicNumber, isExternal, NAMED_IDS, type Particle } from "./particle-tab-helpers";
 
   interface Props {
     selectionState: EntitySelectionState;
@@ -44,10 +44,6 @@
   let highlightedId = $state<number | string | null>(null);
   let showOnlySelected = $state(false);
 
-  function isExternal(p: Particle): p is ExternalOnlyParticle {
-    return typeof p.id === "string";
-  }
-
   function searchText(p: Particle): string {
     if (isExternal(p)) {
       return `${p.localId} ${p.name} ${p.symbol} ${p.label} ext external`;
@@ -72,10 +68,6 @@
 
   // spec: drop electron entirely until ESTAR ships.
   const allBuiltin = $derived(selectionState.allParticles.filter((p) => p.id !== ELECTRON_ID));
-
-  // Named particles (proton Z=1, alpha particle Z=2) sort first by Z, then
-  // remaining builtins by Z, then external-only particles by Z.
-  const NAMED_IDS = new Set([1, 2]);
 
   const flatList = $derived.by<Particle[]>(() => {
     const named = allBuiltin
@@ -145,10 +137,6 @@
     return multiIds.includes(p.id);
   }
 
-  function isAnchor(p: Particle): boolean {
-    return multiIds[0] === p.id;
-  }
-
   function handleMultiToggle(p: Particle): void {
     selectionState.toggleMulti("particle", p.id);
   }
@@ -157,15 +145,6 @@
     // Toggle each non-anchor id (anchor cannot be removed)
     const [, ...rest] = multiIds;
     for (const id of rest) selectionState.toggleMulti("particle", id);
-  }
-
-  function atomicNumber(p: Particle): number {
-    return isExternal(p) ? p.Z : (p.id as number);
-  }
-
-  /** Named particles (proton/alpha) get bold emphasis per spec default (b). */
-  function isNamed(p: Particle): boolean {
-    return !isExternal(p) && NAMED_IDS.has(p.id as number);
   }
 
   // Summary bar derived values
@@ -195,31 +174,6 @@
     if (!isAdvancedMode.value && view === "grid") view = "list";
   });
 
-  /**
-   * Map an atomic number to its (row, column) cell in the standard 18-column
-   * periodic table. Lanthanides (57-71) and actinides (89-103) live on rows
-   * 9 / 10 below the main table; row 8 is an intentional gap.
-   */
-  function periodicPosition(z: number): { row: number; col: number } | null {
-    if (z === 1) return { row: 1, col: 1 };
-    if (z === 2) return { row: 1, col: 18 };
-    if (z >= 3 && z <= 4) return { row: 2, col: z - 2 };
-    if (z >= 5 && z <= 10) return { row: 2, col: z + 8 };
-    if (z >= 11 && z <= 12) return { row: 3, col: z - 10 };
-    if (z >= 13 && z <= 18) return { row: 3, col: z };
-    if (z >= 19 && z <= 36) return { row: 4, col: z - 18 };
-    if (z >= 37 && z <= 54) return { row: 5, col: z - 36 };
-    if (z === 55) return { row: 6, col: 1 };
-    if (z === 56) return { row: 6, col: 2 };
-    if (z >= 57 && z <= 71) return { row: 9, col: z - 53 };
-    if (z >= 72 && z <= 86) return { row: 6, col: z - 68 };
-    if (z === 87) return { row: 7, col: 1 };
-    if (z === 88) return { row: 7, col: 2 };
-    if (z >= 89 && z <= 103) return { row: 10, col: z - 85 };
-    if (z >= 104 && z <= 118) return { row: 7, col: z - 100 };
-    return null;
-  }
-
   const filteredBuiltin = $derived(filteredFlat.filter((p) => !isExternal(p)));
 
   const filteredExternal = $derived(
@@ -228,7 +182,7 @@
       .filter((p) =>
         showOnlySelected ? (isMultiMode ? isMultiSelected(p) : selected?.id === p.id) : true,
       )
-      .sort((a, b) => a.Z - b.Z),
+      .sort((a, b) => a.Z - b.Z) as ExternalOnlyParticle[],
   );
 </script>
 
@@ -288,192 +242,28 @@
   {/if}
 
   {#if view === "grid"}
-    <div class="overflow-x-auto">
-      <div
-        class="grid gap-[2px]"
-        style="grid-template-columns: repeat(18, minmax(0, 1fr)); grid-template-rows: repeat(7, auto) 0.35rem repeat(2, auto); min-width: 360px;"
-        role="listbox"
-        aria-label="Particles (periodic table)"
-        aria-multiselectable={isMultiMode}
-        data-testid="picker-particle-grid"
-      >
-        {#each allBuiltin as p (p.id)}
-          {@const z = atomicNumber(p)}
-          {@const pos = periodicPosition(z)}
-          {@const isMatched = filteredBuiltin.includes(p)}
-          {@const available = isAvailable(p)}
-          {@const inMulti = isMultiSelected(p)}
-          {@const anchor = isAnchor(p)}
-          {@const isSingleSelected = !isMultiMode && selected?.id === p.id}
-          {@const isChecked = isMultiMode ? inMulti : isSingleSelected}
-          {@const isHighlighted = highlightedId === p.id}
-          {@const sym = !isExternal(p) ? (p as ParticleEntity).symbol || "?" : "?"}
-          {#if pos}
-            <button
-              type="button"
-              role="option"
-              aria-selected={isMultiMode ? inMulti : isSingleSelected}
-              aria-disabled={!available || !isMatched || (isMultiMode && anchor)}
-              aria-label="{getParticleListLabel(p, z)}{available ? '' : ' (unavailable)'}"
-              title={getParticleListLabel(p, z)}
-              data-testid="picker-particle-tile-{p.id}"
-              data-available={available ? "1" : "0"}
-              disabled={!available || !isMatched || (isMultiMode && anchor)}
-              style="grid-row: {pos.row}; grid-column: {pos.col};"
-              class={cn(
-                "relative flex aspect-square flex-col items-center justify-center rounded-sm border bg-card p-0 text-center leading-none transition-colors overflow-hidden",
-                available && isMatched ? "hover:bg-accent cursor-pointer" : "pointer-events-none",
-                !available ? "opacity-40" : !isMatched ? "opacity-30" : null,
-                isChecked && "ring-2 ring-inset ring-orange-400 bg-orange-50/60",
-                !isChecked && isHighlighted && available && isMatched && "bg-accent",
-              )}
-              onclick={() => {
-                if (!available || !isMatched) return;
-                if (isMultiMode) {
-                  if (!anchor) handleMultiToggle(p);
-                } else {
-                  onSelect(p);
-                }
-              }}
-            >
-              <span
-                class={cn(
-                  "absolute left-[1px] top-[1px] font-mono text-[min(0.5rem,1.5vw)] leading-none transition-opacity",
-                  isMatched
-                    ? "text-muted-foreground opacity-70"
-                    : "opacity-10 text-muted-foreground",
-                )}>{z}</span
-              >
-              <span
-                class={cn(
-                  "font-mono text-[clamp(11px,1.5vw,18px)] font-bold leading-none transition-opacity",
-                  !isMatched && "opacity-10",
-                )}>{sym}</span
-              >
-            </button>
-          {/if}
-        {/each}
-        <!-- Indicator cells pointing to the lanthanide / actinide rows below. -->
-        <div
-          aria-hidden="true"
-          class="flex aspect-square items-center justify-center rounded-sm border border-dashed bg-muted/30 text-[7px] text-muted-foreground font-mono"
-          style="grid-row: 6; grid-column: 3;"
-        >
-          57-71
-        </div>
-        <div
-          aria-hidden="true"
-          class="flex aspect-square items-center justify-center rounded-sm border border-dashed bg-muted/30 text-[7px] text-muted-foreground font-mono"
-          style="grid-row: 7; grid-column: 3;"
-        >
-          89-103
-        </div>
-      </div>
-    </div>
-    {#if filteredBuiltin.length === 0 && filteredExternal.length === 0}
-      <div class="px-2 py-4 text-center text-sm text-muted-foreground">No particles match.</div>
-    {/if}
-
-    {#if filteredExternal.length > 0}
-      <div class="mt-2 border-t pt-2" data-testid="picker-particle-grid-external">
-        <div class="mb-1 text-[10px] uppercase tracking-wider text-muted-foreground">External</div>
-        <div class="flex flex-wrap gap-1">
-          {#each filteredExternal as p (p.id)}
-            {@const available = isAvailable(p)}
-            {@const inMulti = isMultiSelected(p)}
-            {@const anchor = isAnchor(p)}
-            {@const isSingleSelected = !isMultiMode && selected?.id === p.id}
-            {@const isChecked = isMultiMode ? inMulti : isSingleSelected}
-            <button
-              type="button"
-              role="option"
-              aria-selected={isMultiMode ? inMulti : isSingleSelected}
-              aria-disabled={!available || (isMultiMode && anchor)}
-              data-testid="picker-particle-ext-tile-{p.id}"
-              data-available={available ? "1" : "0"}
-              disabled={!available || (isMultiMode && anchor)}
-              class={cn(
-                "inline-flex items-center gap-1 rounded-sm border bg-card px-1.5 py-0.5 text-[11px]",
-                available ? "hover:bg-accent cursor-pointer" : "opacity-40 pointer-events-none",
-                isChecked && "ring-2 ring-inset ring-orange-400 bg-orange-50/60 font-semibold",
-              )}
-              onclick={() => {
-                if (!available) return;
-                if (isMultiMode) {
-                  if (!anchor) handleMultiToggle(p);
-                } else {
-                  onSelect(p);
-                }
-              }}
-            >
-              <span aria-hidden="true">🔗</span>
-              <span class="font-mono">{p.symbol || p.name}</span>
-              <span class="text-[10px] text-muted-foreground font-mono">Z={p.Z}</span>
-            </button>
-          {/each}
-        </div>
-      </div>
-    {/if}
+    <ParticleGridView
+      {allBuiltin}
+      {filteredBuiltin}
+      {filteredExternal}
+      {selected}
+      {multiIds}
+      {isMultiMode}
+      {highlightedId}
+      {isAvailable}
+      onSelect={(p) => onSelect(p)}
+      onMultiToggle={handleMultiToggle}
+    />
   {:else}
-    <ul
-      role="listbox"
-      aria-label="Particles"
-      aria-multiselectable={isMultiMode}
-      tabindex="0"
-      class="max-h-52 overflow-auto space-y-0.5"
-      data-testid="picker-particle-list"
-    >
-      {#each filteredFlat as p (p.id)}
-        {@const available = isAvailable(p)}
-        {@const inMulti = isMultiSelected(p)}
-        {@const anchor = isAnchor(p)}
-        {@const isSingleSelected = !isMultiMode && selected?.id === p.id}
-        {@const isChecked = isMultiMode ? inMulti : isSingleSelected}
-        {@const isHighlighted = highlightedId === p.id}
-        {@const external = isExternal(p)}
-        {@const z = atomicNumber(p)}
-        {@const named = isNamed(p)}
-        <li role="presentation">
-          <button
-            type="button"
-            role="option"
-            aria-selected={isMultiMode ? inMulti : isSingleSelected}
-            aria-disabled={!available || (isMultiMode && anchor)}
-            data-testid="picker-particle-item-{p.id}"
-            tabindex={-1}
-            disabled={!available || (isMultiMode && anchor)}
-            class={cn(
-              "flex w-full items-center gap-2 rounded px-2 py-1.5 text-sm text-left",
-              available ? "hover:bg-accent cursor-pointer" : "opacity-40 pointer-events-none",
-              isChecked && "ring-1 ring-inset ring-orange-400 bg-orange-50/60 font-semibold",
-              !isChecked && isHighlighted && available && "bg-accent",
-              !isChecked && named && "font-semibold",
-            )}
-            onclick={() => {
-              if (!available) return;
-              if (isMultiMode) {
-                if (!anchor) handleMultiToggle(p);
-              } else {
-                onSelect(p);
-              }
-            }}
-          >
-            <!-- Selection indicator: ✓ / ○ / empty -->
-            <span
-              aria-hidden="true"
-              class="w-4 shrink-0 text-center text-xs {isChecked
-                ? 'font-bold text-orange-700'
-                : 'text-muted-foreground'}">{isChecked ? "✓" : isMultiMode ? "○" : ""}</span
-            >
-            {#if external}<span aria-hidden="true">🔗</span>{/if}
-            {#if named}<span aria-hidden="true" class="mr-0.5">★</span>{/if}
-            <span class="flex-1">{getParticleListLabel(p, z)}</span>
-          </button>
-        </li>
-      {/each}
-      {#if filteredFlat.length === 0}
-        <li class="px-2 py-4 text-center text-sm text-muted-foreground">No particles match.</li>
-      {/if}
-    </ul>
+    <ParticleListView
+      items={filteredFlat}
+      {selected}
+      {multiIds}
+      {isMultiMode}
+      {highlightedId}
+      {isAvailable}
+      onSelect={(p) => onSelect(p)}
+      onMultiToggle={handleMultiToggle}
+    />
   {/if}
 </div>
